DRYing in ServiceToken classes
