+++ src/service-token/node/validator.js
@@ -1,14 +1,10 @@
 'use strict'
 
 const param = require('jquery-param')
-const Deserializer = require('../../jsonapi-serializer/deserializer')
-const { isString } = require('../../utils/typecheck')
-const signedFetch = require('../../utils/signed-fetch')
-
-const deserializer = new Deserializer({ keyForAttribute: 'camelCase' })
+const ServiceTokenBase = require('./base')
 
 /**
- * @class ServiceTokenClient
+ * @class ServiceTokenValidator
  * @classdesc Helper class to request and manage a valid service token in a NodeJS environment.
  *
  * @param {object} [options] the options
@@ -16,24 +12,7 @@ const deserializer = new Deserializer({ keyForAttribute: 'camelCase' })
  *   - secretKey {string}
  *   - baseUrl {string}
  */
-class ServiceTokenValidator {
-  constructor (options = {}) {
-    this.accessKey = options.accessKey
-    this.secretKey = options.secretKey
-    this.digest = options.digest || 'sha1'
-    this.baseUrl = options.baseUrl
-
-    if (!isString(this.baseUrl)) {
-      throw new Error('Key `baseUrl` in `options` parameter is missing or not a string')
-    }
-    if (!isString(this.accessKey)) {
-      throw new Error('Key `accessKey` in `options` parameter is missing or not a string')
-    }
-    if (!isString(this.secretKey)) {
-      throw new Error('Key `secretKey` in `options` parameter is missing or not a string')
-    }
-  }
-
+class ServiceTokenValidator extends ServiceTokenBase {
   /**
    * Verifies if a service token is still valid.
    *
@@ -50,22 +29,6 @@ class ServiceTokenValidator {
     }, options)
 
     return this.__request(url, requestOptions)
-  }
-
-  __request (url, options = {}) {
-    const headers = Object.assign({
-      'Content-Type': 'application/vnd.api+json'
-    }, options.headers || {})
-    const requestOptions = Object.assign({}, options, {
-      accessKey: this.accessKey,
-      secretKey: this.secretKey,
-      digest: this.digest,
-      headers: headers
-    })
-
-    return signedFetch(url, requestOptions).then(json => {
-      return deserializer.deserialize(json)
-    })
   }
 }
